Ensure /var/lib/apt/lists/partial is removed from final image
The file is being correctly removed but is then re-added with a new
timestamp by the 'apt-get clean' command.

When this file is left inside the final image it confuses the
cache_valid_time option of the apt module in ansible - this module will
think that a recent cache is present but actually there is no cache
present because we have removed it.